# Lab4-TodoListRabbitMQ/docs/openapi.yaml
openapi: 3.0.3
info:
  title: TodoList API
  version: 2.0.0
  description: REST API с отдельными обработчиками v1 и v2 для регистрации пользователей, JWT-аутентификации и управления личными задачами.
servers:
  - url: http://localhost:8080
    description: Локальный сервер
...
paths:
...
  /api/{version}/tasks:
    parameters: [{$ref: '#/components/parameters/Version'}]
    get:
      tags: [Tasks]
      security: [{bearerAuth: []}]
      summary: Получить список задач
      description: Возвращает только задачи текущего пользователя. В v2 поддерживаются пагинация offset/limit и выборка полей через include. Поле priority доступно в v2.
      parameters:
        - {$ref: '#/components/parameters/Offset'}
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Include'}
      responses:
        '200':
          description: Список задач
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TaskList'}
              example: [{id: task-001, title: Подготовить отчёт, description: Финальная проверка, status: in_progress, priority: high, created_user_id: usr-001}]
              '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '429': {$ref: '#/components/responses/RateLimited'}
    post:
      tags: [Tasks]
      security: [{bearerAuth: []}]
      summary: Создать задачу
      description: Создаёт задачу текущего пользователя. Для защиты от дублей можно передать Idempotency-Key.
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TaskInputV2'}
            example: {title: Подготовить отчёт, description: Финальная проверка, status: new, priority: high}
      responses:
        '201':
          description: Задача создана
          content: {application/json: {schema: {$ref: '#/components/schemas/Task'}, example: {id: task-001, title: Подготовить отчёт, description: Финальная проверка, status: new, priority: high, created_user_id: usr-001}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '429': {$ref: '#/components/responses/RateLimited'}
  /api/{version}/tasks/{id}:
    parameters:
      - {$ref: '#/components/parameters/Version'}
      - {$ref: '#/components/parameters/TaskId'}
    get:
      tags: [Tasks]
      security: [{bearerAuth: []}]
      summary: Получить задачу
      description: Возвращает задачу текущего пользователя. В v2 параметр include позволяет выбрать дополнительные поля description, status и priority; поля id, title и created_user_id возвращаются всегда.
      parameters: [{$ref: '#/components/parameters/Include'}]
      responses:
        '200':
          description: Задача найдена
          content: {application/json: {schema: {$ref: '#/components/schemas/Task'}, example: {id: task-001, title: Подготовить отчёт, description: Финальная проверка, status: in_progress, priority: high, created_user_id: usr-001}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
    put:
      tags: [Tasks]
      security: [{bearerAuth: []}]
      summary: Обновить задачу
      description: Заменяет данные задачи текущего пользователя.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TaskInputV2'}
            example: {title: Подготовить отчёт, description: Отправить до пятницы, status: done, priority: high}
      responses:
        '200':
          description: Задача обновлена
          content: {application/json: {schema: {$ref: '#/components/schemas/Task'}, example: {id: task-001, title: Подготовить отчёт, description: Отправить до пятницы, status: done, priority: high, created_user_id: usr-001}}}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      tags: [Tasks]
      security: [{bearerAuth: []}]
      summary: Удалить задачу
      description: Удаляет задачу текущего пользователя.
      responses:
        '204': {description: 'Задача удалена; тело ответа отсутствует'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
    serviceToken:
      type: apiKey
      in: header
      name: X-Service-Token
  parameters:
    Version:
      name: version
      in: path
      required: true
      description: Версия API. Используйте v1 или v2.
      schema: {type: string, enum: [v1, v2]}
...
    TaskId:
      name: id
      in: path
      required: true
      description: Идентификатор задачи
      schema: {type: string, example: task-001}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Уникальный ключ запроса для предотвращения создания дублей.
      schema: {type: string, minLength: 1, example: 7f4c2d8e-6d4c-4a35-9a6f-123456789abc}
...
    Offset:
      name: offset
      in: query
      required: false
      description: Количество задач, которое нужно пропустить перед выдачей результата.
      schema: {type: integer, minimum: 0, default: 0, example: 0}
    Limit:
      name: limit
      in: query
      required: false
      description: Максимальное количество задач в ответе. Допустимый диапазон от 1 до 100.
      schema: {type: integer, minimum: 1, maximum: 100, default: 10, example: 10}
    Include:
      name: include
      in: query
      required: false
      description: "Параметр v2. Список дополнительных полей через запятую: description, status, priority. Поля id, title и created_user_id возвращаются всегда."
      schema: {type: string, example: description,status,priority}
  schemas:
...
    TaskInput:
      type: object
      required: [title]
      properties:
        title: {type: string, description: Заголовок задачи, example: Подготовить отчёт}
        description: {type: string, description: Подробное описание, example: Финальная проверка}
        status: {type: string, description: Текущий статус задачи, enum: [new, in_progress, done], default: new, example: new}
    TaskInputV2:
      allOf:
        - {$ref: '#/components/schemas/TaskInput'}
        - type: object
          properties:
            priority: {type: string, enum: [low, medium, high], default: medium, example: high}
    Task:
      allOf:
        - {$ref: '#/components/schemas/TaskInputV2'}
        - type: object
          required: [id, created_user_id]
          properties:
            id: {type: string, example: task-001}
            created_user_id: {type: string, example: usr-001}
    TaskList:
      type: array
      items: {$ref: '#/components/schemas/Task'}
...
    Error:
      type: object
      required: [error]
      properties:
        error: {type: string, example: invalid request}
  responses:
    InvalidRequest:
      description: Некорректное тело запроса
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}, example: {error: invalid request}}}
    Unauthorized:
      description: Требуется корректный JWT
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}, example: {error: authentication required}}}
    NotFound:
      description: Ресурс не найден
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}, example: {error: task not found}}}
    RateLimited:
      description: Превышен лимит 60 запросов в минуту
      headers:
        Retry-After: {description: Секунды до следующей попытки, schema: {type: integer}}
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}, example: {error: rate limit exceeded}}}